+++ proofs/application.php
@@ -147,4 +147,155 @@ return static function() {
                 ->same(\ini_get('memory_limit'));
         },
     )->tag(Tag::ci, Tag::local);
+
+    yield test(
+        'BlackBox can stop on first failure',
+        static function($assert) {
+            $io = Collect::new();
+
+            $result = Application::new([])
+                ->displayOutputVia($io)
+                ->displayErrorVia($io)
+                ->usePrinter(Standard::withoutColors())
+                ->stopOnFailure()
+                ->tryToProve(static function() use (&$value) {
+                    yield proof(
+                        'example',
+                        given(Set\Integers::any()),
+                        static fn($assert, $i) => $assert->number($i),
+                    );
+                    yield proof(
+                        'example',
+                        given(Set\Integers::any()),
+                        static fn($assert, $i) => $assert->true(false),
+                    );
+                    yield proof(
+                        'example',
+                        given(Set\Integers::any()),
+                        static fn($assert, $i) => $assert->number($i),
+                    );
+                });
+
+            $assert->false($result->successful());
+            $assert
+                ->string($io->toString())
+                ->contains('Proofs: 2,');
+        },
+    )->tag(Tag::ci, Tag::local);
+
+    yield test(
+        'Application::map()',
+        static function($assert) {
+            $io = Collect::new();
+
+            $result = Application::new([])
+                ->map(
+                    static fn($app) => $app
+                        ->displayOutputVia($io)
+                        ->displayErrorVia($io)
+                        ->usePrinter(Standard::withoutColors())
+                        ->stopOnFailure(),
+                )
+                ->tryToProve(static function() use (&$value) {
+                    yield proof(
+                        'example',
+                        given(Set\Integers::any()),
+                        static fn($assert, $i) => $assert->number($i),
+                    );
+                    yield proof(
+                        'example',
+                        given(Set\Integers::any()),
+                        static fn($assert, $i) => $assert->true(false),
+                    );
+                    yield proof(
+                        'example',
+                        given(Set\Integers::any()),
+                        static fn($assert, $i) => $assert->number($i),
+                    );
+                });
+
+            $assert->false($result->successful());
+            $assert
+                ->string($io->toString())
+                ->contains('Proofs: 2,');
+        },
+    )->tag(Tag::ci, Tag::local);
+
+    yield test(
+        'Application::when() enabled',
+        static function($assert) {
+            $io = Collect::new();
+
+            $result = Application::new([])
+                ->displayOutputVia($io)
+                ->displayErrorVia($io)
+                ->usePrinter(Standard::withoutColors())
+                ->when(
+                    true,
+                    static fn($app) => $app
+                        ->stopOnFailure(),
+                )
+                ->tryToProve(static function() use (&$value) {
+                    yield proof(
+                        'example',
+                        given(Set\Integers::any()),
+                        static fn($assert, $i) => $assert->number($i),
+                    );
+                    yield proof(
+                        'example',
+                        given(Set\Integers::any()),
+                        static fn($assert, $i) => $assert->true(false),
+                    );
+                    yield proof(
+                        'example',
+                        given(Set\Integers::any()),
+                        static fn($assert, $i) => $assert->number($i),
+                    );
+                });
+
+            $assert->false($result->successful());
+            $assert
+                ->string($io->toString())
+                ->contains('Proofs: 2,');
+        },
+    )->tag(Tag::ci, Tag::local);
+
+    yield test(
+        'Application::when() disabled',
+        static function($assert) {
+            $io = Collect::new();
+
+            $result = Application::new([])
+                ->displayOutputVia($io)
+                ->displayErrorVia($io)
+                ->usePrinter(Standard::withoutColors())
+                ->when(
+                    false,
+                    static fn($app) => $app
+                        ->stopOnFailure(),
+                )
+                ->tryToProve(static function() use (&$value) {
+                    yield proof(
+                        'example',
+                        given(Set\Integers::any()),
+                        static fn($assert, $i) => $assert->number($i),
+                    );
+                    yield proof(
+                        'example',
+                        given(Set\Integers::any()),
+                        static fn($assert, $i) => $assert->true(false),
+                    );
+                    yield proof(
+                        'example',
+                        given(Set\Integers::any()),
+                        static fn($assert, $i) => $assert->number($i),
+                    );
+                });
+
+            $assert->false($result->successful());
+            $assert
+                ->string($io->toString())
+                ->contains('Proofs: 3,');
+        },
+    )->tag(Tag::ci, Tag::local);
 };